initial commit for ISMIR 2020

+++ audio_conditioned_unet/video_utils.py
@@ -5,8 +5,8 @@ import os
 import matplotlib.cm as cm
 import matplotlib.pyplot as plt
 import numpy as np
-import soundfile as sf
 
+from audio_conditioned_unet.utils import render_audio
 from matplotlib.colors import ListedColormap
 
 
@@ -41,71 +41,28 @@ def mux_video_audio(path_video, path_audio, path_output='output_audio.mp4'):
     check_call(["ffmpeg", "-y", "-i", path_video, "-i", path_audio, "-shortest", path_output])
 
 
-def fluidsynth(midi, fs=44100, sf2_path=None):
-    """Synthesize using fluidsynth.
-    Copied and adapted from `pretty_midi`.
+def create_video(observation_images, midi_path, piece_name, spectrogram_params, sf_path, path="../videos", real_perf=False):
 
-    Parameters
-    ----------
-    midi : pm.PrettyMidi
-    fs : int
-        Sampling rate to synthesize at.
-    sf2_path : str
-        Path to a .sf2 file.
-        Default ``None``, which uses the TimGM6mb.sf2 file included with
-        ``pretty_midi``.
-    Returns
-    -------
-    synthesized : np.ndarray
-        Waveform of the MIDI data, synthesized at ``fs``.
-    """
-    # If there are no instruments, or all instruments have no notes, return
-    # an empty array
-    if len(midi.instruments) == 0 or all(len(i.notes) == 0 for i in midi.instruments):
-        return np.array([])
-    # Get synthesized waveform for each instrument
-    waveforms = []
-    for i in midi.instruments:
-        if len(i.notes) > 0:
-            waveforms.append(i.fluidsynth(fs=fs, sf2_path=sf2_path))
+    if not os.path.exists(path):
+        os.mkdir(path)
 
-    # Allocate output waveform, with #sample = max length of all waveforms
-    synthesized = np.zeros(np.max([w.shape[0] for w in waveforms]))
-
-    # Sum all waveforms in
-    for waveform in waveforms:
-        synthesized[:waveform.shape[0]] += waveform
-
-    # Scale to [-1, 1]
-    synthesized /= 2**16
-    # synthesized = synthesized.astype(np.int16)
-
-    # normalize
-    synthesized /= float(np.max(np.abs(synthesized)))
-
-    return synthesized
-
-
-def create_video(observation_images, midi, piece_name, spectrogram_params, sf_path):
-    fn_audio = '../videos/tmp.wav'
-
-    perf_audio = fluidsynth(midi, fs=spectrogram_params['sample_rate'], sf2_path=sf_path)
-
-    # let it start at the first onset position
-    # perf_audio = midi_synth[int(onsets[0] * spectrogram_params['sample_rate'] / spectrogram_params['fps']):]
-
-    # get synthesized MIDI as WAV
-    sf.write(fn_audio, perf_audio, spectrogram_params['sample_rate'])
+    if real_perf:
+        # midi path will be the wav path
+        fn_audio = midi_path
+    else:
+        fn_audio = render_audio(midi_path, sf_path)
 
     # frame rate video is now based on the piano roll's frame rate
-    path_video = write_video(observation_images, fn_output='../videos/test.mp4',
+    path_video = write_video(observation_images, fn_output=os.path.join(path, 'test.mp4'),
                              frame_rate=spectrogram_params['fps'], overwrite=True)
 
     # mux video and audio with ffmpeg
-    mux_video_audio(path_video, fn_audio, path_output='../videos/{}.mp4'.format(piece_name))
+    mux_video_audio(path_video, fn_audio, path_output=os.path.join(path, '{}.mp4'.format(piece_name)))
 
     # clean up
-    os.remove(fn_audio)
+    if not real_perf:
+        os.remove(fn_audio)
+
     os.remove(path_video)
 
 
@@ -113,9 +70,10 @@ def prepare_score_for_render(score, mask, cmap=None):
 
     if cmap is None:
         cmap = get_transparent_cmap(plt.get_cmap('YlOrBr'), alpha=0.5, n_steps_blend=50)
-        # cmap = get_transparent_cmap(plt.get_cmap('inferno'), alpha=0.5, n_steps_blend=50)
 
-    mask = cv2.resize(mask, (score.shape[1], score.shape[0]), interpolation=cv2.INTER_NEAREST)
+    if mask.shape[0] != score.shape[0] or mask.shape[1] != score.shape[1]:
+
+        mask = cv2.resize(mask, (score.shape[1], score.shape[0]), interpolation=cv2.INTER_NEAREST)
 
     mask = np.array(cmap(mask), dtype=np.float32)
 
@@ -125,7 +83,7 @@ def prepare_score_for_render(score, mask, cmap=None):
     indices = mask[:, :, 3] > 0.
     img[indices] = cv2.addWeighted(img[indices], 0.5, mask[indices], 0.5, 0)
 
-    return img[:, :, :3]
+    return img[:, :, :3], mask
 
 
 def prepare_spec_for_render(spec, score, scale_factor=5):
